Document interfaces - reorder implementations to match interfaces.
