+++ Source/Services/Callbacks/MetricsCollector.cs
@@ -15,22 +15,37 @@ namespace Dolittle.Runtime.Services.Callbacks
     [Singleton]
     public class MetricsCollector : ICanProvideMetrics, IMetricsCollector
     {
+        Counter _totalCallbacksRegistered;
+        Counter _totalCallbacksCalled;
         Counter _totalCallbackTime;
+        Counter _totalCallbacksFailed;
+        Counter _totalCallbacksUnregistered;
         Counter _totalSchedulesMissed;
         Counter _totalSchedulesMissedTime;
         Counter _totalCallbackLoopsFailed;
-        Counter _totalCallbacksCalled;
-        Counter _totalCallbacksFailed;
-        Counter _totalCallbacksRegistered;
-        Counter _totalCallbacksUnregistered;
-
 
         /// <inheritdoc/>
         public IEnumerable<Collector> Provide(IMetricFactory metricFactory)
         {
+            _totalCallbacksRegistered = metricFactory.Counter(
+                "dolittle_system_runtime_services_callbacks_registered_total",
+                "Callbacks total number of registered callbacks");
+
+            _totalCallbacksCalled = metricFactory.Counter(
+                "dolittle_system_runtime_services_callbacks_calls_total",
+                "Callbacks total number of called callbacks");
+
             _totalCallbackTime = metricFactory.Counter(
                 "dolittle_system_runtime_services_callbacks_time_seconds_total",
                 "Callbacks total time spent calling callbacks");
+
+            _totalCallbacksFailed = metricFactory.Counter(
+                "dolittle_system_runtime_services_callbacks_failed_calls_total",
+                "Callbacks total number of called callbacks that failed");
+
+            _totalCallbacksUnregistered = metricFactory.Counter(
+                "dolittle_system_runtime_services_callbacks_unregistered_total",
+                "Callbacks total number of unregistered callbacks");
 
             _totalSchedulesMissed = metricFactory.Counter(
                 "dolittle_system_runtime_services_callbacks_schedules_missed_total",
@@ -40,42 +55,42 @@ namespace Dolittle.Runtime.Services.Callbacks
                 "dolittle_system_runtime_services_callbacks_schedules_missed_time_seconds_total",
                 "Callbacks total time delays for missed callback schedules");
 
-            _totalCallbacksCalled = metricFactory.Counter(
-                "dolittle_system_runtime_services_callbacks_calls_total",
-                "Callbacks total number of called callbacks");
-
-            _totalCallbacksFailed = metricFactory.Counter(
-                "dolittle_system_runtime_services_callbacks_failed_calls_total",
-                "Callbacks total number of called callbacks that failed");
-
             _totalCallbackLoopsFailed = metricFactory.Counter(
                 "dolittle_system_runtime_services_callbacks_failed_call_loops_total",
                 "Callbacks total number of failed callback loops");
 
-            _totalCallbacksRegistered = metricFactory.Counter(
-                "dolittle_system_runtime_services_callbacks_registered_total",
-                "Callbacks total number of registered callbacks");
-
-            _totalCallbacksUnregistered = metricFactory.Counter(
-                "dolittle_system_runtime_services_callbacks_unregistered_total",
-                "Callbacks total number of unregistered callbacks");
-
             return new Collector[]
             {
+                _totalCallbacksRegistered,
+                _totalCallbacksCalled,
                 _totalCallbackTime,
+                _totalCallbacksFailed,
+                _totalCallbacksUnregistered,
                 _totalSchedulesMissed,
                 _totalSchedulesMissedTime,
-                _totalCallbacksCalled,
-                _totalCallbacksFailed,
                 _totalCallbackLoopsFailed,
-                _totalCallbacksRegistered,
-                _totalCallbacksUnregistered,
             };
         }
 
         /// <inheritdoc/>
+        public void IncrementTotalCallbacksRegistered()
+            => _totalCallbacksRegistered.Inc();
+
+        /// <inheritdoc/>
+        public void IncrementTotalCallbacksCalled()
+            => _totalCallbacksCalled.Inc();
+
+        /// <inheritdoc/>
         public void AddToTotalCallbackTime(TimeSpan elapsed)
             => _totalCallbackTime.Inc(elapsed.TotalSeconds);
+
+        /// <inheritdoc/>
+        public void IncrementTotalCallbacksFailed()
+            => _totalCallbacksFailed.Inc();
+
+        /// <inheritdoc/>
+        public void IncrementTotalCallbacksUnregistered()
+            => _totalCallbacksUnregistered.Inc();
 
         /// <inheritdoc/>
         public void IncrementTotalSchedulesMissed()
@@ -88,21 +103,5 @@ namespace Dolittle.Runtime.Services.Callbacks
         /// <inheritdoc/>
         public void IncrementTotalCallbackLoopsFailed()
             => _totalCallbackLoopsFailed.Inc();
-
-        /// <inheritdoc/>
-        public void IncrementTotalCallbacksCalled()
-            => _totalCallbacksCalled.Inc();
-
-        /// <inheritdoc/>
-        public void IncrementTotalCallbacksFailed()
-            => _totalCallbacksFailed.Inc();
-
-        /// <inheritdoc/>
-        public void IncrementTotalCallbacksRegistered()
-            => _totalCallbacksRegistered.Inc();
-
-        /// <inheritdoc/>
-        public void IncrementTotalCallbacksUnregistered()
-            => _totalCallbacksUnregistered.Inc();
     }
 }
